+++ cognite/experimental/_client.py
@@ -13,6 +13,7 @@ from cognite.experimental._api.model_hosting import ModelHostingAPI
 from cognite.experimental._api.pnid_parsing import PNIDParsingAPI
 from cognite.experimental._api.relationships import RelationshipsAPI
 from cognite.experimental._api.resource_typing import ResourceTypingAPI
+from cognite.experimental._api.schema_completion import SchemaCompletionAPI
 from cognite.experimental._api.types import TypesAPI
 from cognite.experimental._api.unstructured import GrepAPI
 
@@ -94,4 +95,6 @@ class CogniteClient(Client):
         self.entity_extraction = EntityExtractionAPI(self._config, api_version="playground", cognite_client=self)
         self.pnid_parsing = PNIDParsingAPI(self._config, api_version="playground", cognite_client=self)
         self.resource_typing = ResourceTypingAPI(self._config, api_version="playground", cognite_client=self)
+        self.schemas = SchemaCompletionAPI(self._config, api_version="playground", cognite_client=self)
+
         self.functions = FunctionsAPI(self.config, api_version="playground", cognite_client=self)