+++ CHANGELOG.md
@@ -12,7 +12,13 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## [0.12.0] - 2020-06-30
+## [0.12.1] - 2020-07-07
+### Added
+- The entity matching methods now have additional options for classifier and feature types.
+- PNID detect patterns endpoint added.
+- Schema completion endpoint added.
+
+## [0.12.0] - 2020-07-07
 ### Changed
 - Synthetic time series and labels removed, as they were moved to non-experimental status in the main SDK.